refactor: move card to own file
