+++ src/@types/types.d.ts
@@ -3,3 +3,11 @@ type SideNavRoute = {
     text: string;
     icon: string;
 }
+
+type AlgorithmData = {
+    name: string;
+    displayAlgorithm: string;
+    faces: string;
+    edges: string;
+    type: string;
+};
